Se emparejaron las entidades

+++ src/main/java/com/backend/digitalhouse/ClinicaOdontologica/dto/entrada/modificado/OdontologoModificacionEntradaDto.java
@@ -6,7 +6,7 @@ import javax.validation.constraints.NotNull;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class OdontologoModificacionEntradaDto {
     @NotNull
-    private int id;
+    private long id;
 
     @NotNull
     private String matricula;
@@ -21,14 +21,14 @@ public class OdontologoModificacionEntradaDto {
     public OdontologoModificacionEntradaDto() {
     }
 
-    public OdontologoModificacionEntradaDto(int id, String matricula, String nombre, String apellido) {
+    public OdontologoModificacionEntradaDto(long id, String matricula, String nombre, String apellido) {
         this.id = id;
         this.matricula = matricula;
         this.nombre = nombre;
         this.apellido = apellido;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
